Establish native tracing equivalence contract
