+++ tailtriage-tracing/tests/expected/equivalence/duration_only_legacy.run.json
@@ -1,0 +1,189 @@
+{
+  "service_name": "equivalence-service",
+  "mode": "light",
+  "effective_core_config": {
+    "mode": "light",
+    "capture_limits": {
+      "max_requests": 100000,
+      "max_stages": 200000,
+      "max_queues": 200000,
+      "max_inflight_snapshots": 200000,
+      "max_runtime_snapshots": 100000
+    },
+    "strict_lifecycle": false
+  },
+  "requests": [
+    {
+      "request_id": "legacy-1",
+      "route": "/legacy",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 100000,
+      "started_at_run_us": null,
+      "finished_at_run_us": null
+    },
+    {
+      "request_id": "legacy-2",
+      "route": "/legacy",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 100000,
+      "started_at_run_us": null,
+      "finished_at_run_us": null
+    },
+    {
+      "request_id": "legacy-3",
+      "route": "/legacy",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 100000,
+      "started_at_run_us": null,
+      "finished_at_run_us": null
+    },
+    {
+      "request_id": "legacy-4",
+      "route": "/legacy",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 100000,
+      "started_at_run_us": null,
+      "finished_at_run_us": null
+    },
+    {
+      "request_id": "legacy-5",
+      "route": "/legacy",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 100000,
+      "started_at_run_us": null,
+      "finished_at_run_us": null
+    },
+    {
+      "request_id": "legacy-6",
+      "route": "/legacy",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 100000,
+      "started_at_run_us": null,
+      "finished_at_run_us": null
+    }
+  ],
+  "stages": [
+    {
+      "request_id": "legacy-1",
+      "stage": "db",
+      "success": true,
+      "completed": true,
+      "latency_us": 15000,
+      "started_at_run_us": null,
+      "finished_at_run_us": null
+    },
+    {
+      "request_id": "legacy-2",
+      "stage": "db",
+      "success": true,
+      "completed": true,
+      "latency_us": 15000,
+      "started_at_run_us": null,
+      "finished_at_run_us": null
+    },
+    {
+      "request_id": "legacy-3",
+      "stage": "db",
+      "success": true,
+      "completed": true,
+      "latency_us": 15000,
+      "started_at_run_us": null,
+      "finished_at_run_us": null
+    },
+    {
+      "request_id": "legacy-4",
+      "stage": "db",
+      "success": true,
+      "completed": true,
+      "latency_us": 15000,
+      "started_at_run_us": null,
+      "finished_at_run_us": null
+    },
+    {
+      "request_id": "legacy-5",
+      "stage": "db",
+      "success": true,
+      "completed": true,
+      "latency_us": 15000,
+      "started_at_run_us": null,
+      "finished_at_run_us": null
+    },
+    {
+      "request_id": "legacy-6",
+      "stage": "db",
+      "success": true,
+      "completed": true,
+      "latency_us": 15000,
+      "started_at_run_us": null,
+      "finished_at_run_us": null
+    }
+  ],
+  "queues": [
+    {
+      "request_id": "legacy-1",
+      "queue": "permits",
+      "depth_at_start": 3,
+      "completed": true,
+      "wait_us": 65000,
+      "waited_from_run_us": null,
+      "waited_until_run_us": null
+    },
+    {
+      "request_id": "legacy-2",
+      "queue": "permits",
+      "depth_at_start": 3,
+      "completed": true,
+      "wait_us": 65000,
+      "waited_from_run_us": null,
+      "waited_until_run_us": null
+    },
+    {
+      "request_id": "legacy-3",
+      "queue": "permits",
+      "depth_at_start": 3,
+      "completed": true,
+      "wait_us": 65000,
+      "waited_from_run_us": null,
+      "waited_until_run_us": null
+    },
+    {
+      "request_id": "legacy-4",
+      "queue": "permits",
+      "depth_at_start": 3,
+      "completed": true,
+      "wait_us": 65000,
+      "waited_from_run_us": null,
+      "waited_until_run_us": null
+    },
+    {
+      "request_id": "legacy-5",
+      "queue": "permits",
+      "depth_at_start": 3,
+      "completed": true,
+      "wait_us": 65000,
+      "waited_from_run_us": null,
+      "waited_until_run_us": null
+    },
+    {
+      "request_id": "legacy-6",
+      "queue": "permits",
+      "depth_at_start": 3,
+      "completed": true,
+      "wait_us": 65000,
+      "waited_from_run_us": null,
+      "waited_until_run_us": null
+    }
+  ],
+  "semantic_truncation": {
+    "limits_hit": false,
+    "dropped_requests": 0,
+    "dropped_stages": 0,
+    "dropped_queues": 0
+  }
+}
