+++ tailtriage-tracing/tests/expected/equivalence/precise_route_divergent.run.json
@@ -1,0 +1,243 @@
+{
+  "service_name": "equivalence-service",
+  "mode": "light",
+  "effective_core_config": {
+    "mode": "light",
+    "capture_limits": {
+      "max_requests": 100000,
+      "max_stages": 200000,
+      "max_queues": 200000,
+      "max_inflight_snapshots": 200000,
+      "max_runtime_snapshots": 100000
+    },
+    "strict_lifecycle": false
+  },
+  "requests": [
+    {
+      "request_id": "route-1",
+      "route": "/queued",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 100000,
+      "started_at_run_us": 0,
+      "finished_at_run_us": 100000
+    },
+    {
+      "request_id": "route-2",
+      "route": "/queued",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 100000,
+      "started_at_run_us": 200000,
+      "finished_at_run_us": 300000
+    },
+    {
+      "request_id": "route-3",
+      "route": "/queued",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 100000,
+      "started_at_run_us": 400000,
+      "finished_at_run_us": 500000
+    },
+    {
+      "request_id": "route-4",
+      "route": "/queued",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 100000,
+      "started_at_run_us": 600000,
+      "finished_at_run_us": 700000
+    },
+    {
+      "request_id": "route-5",
+      "route": "/downstream",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 100000,
+      "started_at_run_us": 800000,
+      "finished_at_run_us": 900000
+    },
+    {
+      "request_id": "route-6",
+      "route": "/downstream",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 100000,
+      "started_at_run_us": 1000000,
+      "finished_at_run_us": 1100000
+    },
+    {
+      "request_id": "route-7",
+      "route": "/downstream",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 100000,
+      "started_at_run_us": 1200000,
+      "finished_at_run_us": 1300000
+    },
+    {
+      "request_id": "route-8",
+      "route": "/downstream",
+      "kind": null,
+      "outcome": "error",
+      "latency_us": 100000,
+      "started_at_run_us": 1400000,
+      "finished_at_run_us": 1500000
+    }
+  ],
+  "stages": [
+    {
+      "request_id": "route-1",
+      "stage": "db",
+      "success": true,
+      "completed": true,
+      "latency_us": 10000,
+      "started_at_run_us": 10000,
+      "finished_at_run_us": 20000
+    },
+    {
+      "request_id": "route-2",
+      "stage": "db",
+      "success": true,
+      "completed": true,
+      "latency_us": 10000,
+      "started_at_run_us": 210000,
+      "finished_at_run_us": 220000
+    },
+    {
+      "request_id": "route-3",
+      "stage": "db",
+      "success": true,
+      "completed": true,
+      "latency_us": 10000,
+      "started_at_run_us": 410000,
+      "finished_at_run_us": 420000
+    },
+    {
+      "request_id": "route-4",
+      "stage": "db",
+      "success": true,
+      "completed": true,
+      "latency_us": 10000,
+      "started_at_run_us": 610000,
+      "finished_at_run_us": 620000
+    },
+    {
+      "request_id": "route-5",
+      "stage": "db",
+      "success": true,
+      "completed": true,
+      "latency_us": 75000,
+      "started_at_run_us": 810000,
+      "finished_at_run_us": 885000
+    },
+    {
+      "request_id": "route-6",
+      "stage": "db",
+      "success": true,
+      "completed": true,
+      "latency_us": 75000,
+      "started_at_run_us": 1010000,
+      "finished_at_run_us": 1085000
+    },
+    {
+      "request_id": "route-7",
+      "stage": "db",
+      "success": false,
+      "completed": true,
+      "latency_us": 75000,
+      "started_at_run_us": 1210000,
+      "finished_at_run_us": 1285000
+    },
+    {
+      "request_id": "route-8",
+      "stage": "db",
+      "success": true,
+      "completed": true,
+      "latency_us": 75000,
+      "started_at_run_us": 1410000,
+      "finished_at_run_us": 1485000
+    }
+  ],
+  "queues": [
+    {
+      "request_id": "route-1",
+      "queue": "permits",
+      "depth_at_start": 5,
+      "completed": true,
+      "wait_us": 70000,
+      "waited_from_run_us": 1000,
+      "waited_until_run_us": 71000
+    },
+    {
+      "request_id": "route-2",
+      "queue": "permits",
+      "depth_at_start": 5,
+      "completed": true,
+      "wait_us": 70000,
+      "waited_from_run_us": 201000,
+      "waited_until_run_us": 271000
+    },
+    {
+      "request_id": "route-3",
+      "queue": "permits",
+      "depth_at_start": 5,
+      "completed": true,
+      "wait_us": 70000,
+      "waited_from_run_us": 401000,
+      "waited_until_run_us": 471000
+    },
+    {
+      "request_id": "route-4",
+      "queue": "permits",
+      "depth_at_start": 5,
+      "completed": true,
+      "wait_us": 70000,
+      "waited_from_run_us": 601000,
+      "waited_until_run_us": 671000
+    },
+    {
+      "request_id": "route-5",
+      "queue": "permits",
+      "depth_at_start": 5,
+      "completed": true,
+      "wait_us": 5000,
+      "waited_from_run_us": 801000,
+      "waited_until_run_us": 806000
+    },
+    {
+      "request_id": "route-6",
+      "queue": "permits",
+      "depth_at_start": 5,
+      "completed": true,
+      "wait_us": 5000,
+      "waited_from_run_us": 1001000,
+      "waited_until_run_us": 1006000
+    },
+    {
+      "request_id": "route-7",
+      "queue": "permits",
+      "depth_at_start": 5,
+      "completed": true,
+      "wait_us": 5000,
+      "waited_from_run_us": 1201000,
+      "waited_until_run_us": 1206000
+    },
+    {
+      "request_id": "route-8",
+      "queue": "permits",
+      "depth_at_start": 5,
+      "completed": true,
+      "wait_us": 5000,
+      "waited_from_run_us": 1401000,
+      "waited_until_run_us": 1406000
+    }
+  ],
+  "semantic_truncation": {
+    "limits_hit": false,
+    "dropped_requests": 0,
+    "dropped_stages": 0,
+    "dropped_queues": 0
+  }
+}
