+++ tailtriage-tracing/tests/expected/equivalence/semantic_retention_limits.run.json
@@ -1,0 +1,81 @@
+{
+  "service_name": "equivalence-service",
+  "mode": "light",
+  "effective_core_config": {
+    "mode": "light",
+    "capture_limits": {
+      "max_requests": 2,
+      "max_stages": 2,
+      "max_queues": 2,
+      "max_inflight_snapshots": 200000,
+      "max_runtime_snapshots": 100000
+    },
+    "strict_lifecycle": false
+  },
+  "requests": [
+    {
+      "request_id": "limit-1",
+      "route": "/limited",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 80000,
+      "started_at_run_us": 0,
+      "finished_at_run_us": 80000
+    },
+    {
+      "request_id": "limit-2",
+      "route": "/limited",
+      "kind": null,
+      "outcome": "ok",
+      "latency_us": 80000,
+      "started_at_run_us": 100000,
+      "finished_at_run_us": 180000
+    }
+  ],
+  "stages": [
+    {
+      "request_id": "limit-1",
+      "stage": "stage-1",
+      "success": true,
+      "completed": true,
+      "latency_us": 20000,
+      "started_at_run_us": 30000,
+      "finished_at_run_us": 50000
+    },
+    {
+      "request_id": "limit-2",
+      "stage": "stage-2",
+      "success": true,
+      "completed": true,
+      "latency_us": 20000,
+      "started_at_run_us": 130000,
+      "finished_at_run_us": 150000
+    }
+  ],
+  "queues": [
+    {
+      "request_id": "limit-1",
+      "queue": "queue-1",
+      "depth_at_start": 1,
+      "completed": true,
+      "wait_us": 30000,
+      "waited_from_run_us": 1000,
+      "waited_until_run_us": 31000
+    },
+    {
+      "request_id": "limit-2",
+      "queue": "queue-2",
+      "depth_at_start": 2,
+      "completed": true,
+      "wait_us": 30000,
+      "waited_from_run_us": 101000,
+      "waited_until_run_us": 131000
+    }
+  ],
+  "semantic_truncation": {
+    "limits_hit": true,
+    "dropped_requests": 3,
+    "dropped_stages": 3,
+    "dropped_queues": 3
+  }
+}
