+++ tailtriage-tracing/tests/fixtures/equivalence/duration_only_legacy.jsonl
@@ -1,0 +1,18 @@
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000000,"finished_at_unix_ms":1700000000100,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"legacy-1","tt.route":"/legacy","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000020,"finished_at_unix_ms":1700000000035,"duration_us":15000,"fields":{"tt.kind":"stage","tt.request_id":"legacy-1","tt.stage":"db","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000001,"finished_at_unix_ms":1700000000066,"duration_us":65000,"fields":{"tt.kind":"queue","tt.request_id":"legacy-1","tt.queue":"permits","tt.depth_at_start":3}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000120,"finished_at_unix_ms":1700000000220,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"legacy-2","tt.route":"/legacy","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000140,"finished_at_unix_ms":1700000000155,"duration_us":15000,"fields":{"tt.kind":"stage","tt.request_id":"legacy-2","tt.stage":"db","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000121,"finished_at_unix_ms":1700000000186,"duration_us":65000,"fields":{"tt.kind":"queue","tt.request_id":"legacy-2","tt.queue":"permits","tt.depth_at_start":3}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000240,"finished_at_unix_ms":1700000000340,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"legacy-3","tt.route":"/legacy","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000260,"finished_at_unix_ms":1700000000275,"duration_us":15000,"fields":{"tt.kind":"stage","tt.request_id":"legacy-3","tt.stage":"db","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000241,"finished_at_unix_ms":1700000000306,"duration_us":65000,"fields":{"tt.kind":"queue","tt.request_id":"legacy-3","tt.queue":"permits","tt.depth_at_start":3}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000360,"finished_at_unix_ms":1700000000460,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"legacy-4","tt.route":"/legacy","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000380,"finished_at_unix_ms":1700000000395,"duration_us":15000,"fields":{"tt.kind":"stage","tt.request_id":"legacy-4","tt.stage":"db","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000361,"finished_at_unix_ms":1700000000426,"duration_us":65000,"fields":{"tt.kind":"queue","tt.request_id":"legacy-4","tt.queue":"permits","tt.depth_at_start":3}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000480,"finished_at_unix_ms":1700000000580,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"legacy-5","tt.route":"/legacy","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000500,"finished_at_unix_ms":1700000000515,"duration_us":15000,"fields":{"tt.kind":"stage","tt.request_id":"legacy-5","tt.stage":"db","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000481,"finished_at_unix_ms":1700000000546,"duration_us":65000,"fields":{"tt.kind":"queue","tt.request_id":"legacy-5","tt.queue":"permits","tt.depth_at_start":3}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000600,"finished_at_unix_ms":1700000000700,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"legacy-6","tt.route":"/legacy","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000620,"finished_at_unix_ms":1700000000635,"duration_us":15000,"fields":{"tt.kind":"stage","tt.request_id":"legacy-6","tt.stage":"db","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000601,"finished_at_unix_ms":1700000000666,"duration_us":65000,"fields":{"tt.kind":"queue","tt.request_id":"legacy-6","tt.queue":"permits","tt.depth_at_start":3}}}
