+++ tailtriage-tracing/tests/fixtures/equivalence/precise_route_divergent.jsonl
@@ -1,0 +1,24 @@
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000000,"finished_at_unix_ms":1700000000100,"started_at_run_us":0,"finished_at_run_us":100000,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"route-1","tt.route":"/queued","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000010,"finished_at_unix_ms":1700000000020,"started_at_run_us":10000,"finished_at_run_us":20000,"duration_us":10000,"fields":{"tt.kind":"stage","tt.request_id":"route-1","tt.stage":"db","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000001,"finished_at_unix_ms":1700000000071,"started_at_run_us":1000,"finished_at_run_us":71000,"duration_us":70000,"fields":{"tt.kind":"queue","tt.request_id":"route-1","tt.queue":"permits","tt.depth_at_start":5}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000200,"finished_at_unix_ms":1700000000300,"started_at_run_us":200000,"finished_at_run_us":300000,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"route-2","tt.route":"/queued","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000210,"finished_at_unix_ms":1700000000220,"started_at_run_us":210000,"finished_at_run_us":220000,"duration_us":10000,"fields":{"tt.kind":"stage","tt.request_id":"route-2","tt.stage":"db","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000201,"finished_at_unix_ms":1700000000271,"started_at_run_us":201000,"finished_at_run_us":271000,"duration_us":70000,"fields":{"tt.kind":"queue","tt.request_id":"route-2","tt.queue":"permits","tt.depth_at_start":5}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000400,"finished_at_unix_ms":1700000000500,"started_at_run_us":400000,"finished_at_run_us":500000,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"route-3","tt.route":"/queued","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000410,"finished_at_unix_ms":1700000000420,"started_at_run_us":410000,"finished_at_run_us":420000,"duration_us":10000,"fields":{"tt.kind":"stage","tt.request_id":"route-3","tt.stage":"db","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000401,"finished_at_unix_ms":1700000000471,"started_at_run_us":401000,"finished_at_run_us":471000,"duration_us":70000,"fields":{"tt.kind":"queue","tt.request_id":"route-3","tt.queue":"permits","tt.depth_at_start":5}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000600,"finished_at_unix_ms":1700000000700,"started_at_run_us":600000,"finished_at_run_us":700000,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"route-4","tt.route":"/queued","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000610,"finished_at_unix_ms":1700000000620,"started_at_run_us":610000,"finished_at_run_us":620000,"duration_us":10000,"fields":{"tt.kind":"stage","tt.request_id":"route-4","tt.stage":"db","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000601,"finished_at_unix_ms":1700000000671,"started_at_run_us":601000,"finished_at_run_us":671000,"duration_us":70000,"fields":{"tt.kind":"queue","tt.request_id":"route-4","tt.queue":"permits","tt.depth_at_start":5}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000800,"finished_at_unix_ms":1700000000900,"started_at_run_us":800000,"finished_at_run_us":900000,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"route-5","tt.route":"/downstream","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000810,"finished_at_unix_ms":1700000000885,"started_at_run_us":810000,"finished_at_run_us":885000,"duration_us":75000,"fields":{"tt.kind":"stage","tt.request_id":"route-5","tt.stage":"db","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000801,"finished_at_unix_ms":1700000000806,"started_at_run_us":801000,"finished_at_run_us":806000,"duration_us":5000,"fields":{"tt.kind":"queue","tt.request_id":"route-5","tt.queue":"permits","tt.depth_at_start":5}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000001000,"finished_at_unix_ms":1700000001100,"started_at_run_us":1000000,"finished_at_run_us":1100000,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"route-6","tt.route":"/downstream","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000001010,"finished_at_unix_ms":1700000001085,"started_at_run_us":1010000,"finished_at_run_us":1085000,"duration_us":75000,"fields":{"tt.kind":"stage","tt.request_id":"route-6","tt.stage":"db","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000001001,"finished_at_unix_ms":1700000001006,"started_at_run_us":1001000,"finished_at_run_us":1006000,"duration_us":5000,"fields":{"tt.kind":"queue","tt.request_id":"route-6","tt.queue":"permits","tt.depth_at_start":5}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000001200,"finished_at_unix_ms":1700000001300,"started_at_run_us":1200000,"finished_at_run_us":1300000,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"route-7","tt.route":"/downstream","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000001210,"finished_at_unix_ms":1700000001285,"started_at_run_us":1210000,"finished_at_run_us":1285000,"duration_us":75000,"fields":{"tt.kind":"stage","tt.request_id":"route-7","tt.stage":"db","tt.success":false}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000001201,"finished_at_unix_ms":1700000001206,"started_at_run_us":1201000,"finished_at_run_us":1206000,"duration_us":5000,"fields":{"tt.kind":"queue","tt.request_id":"route-7","tt.queue":"permits","tt.depth_at_start":5}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000001400,"finished_at_unix_ms":1700000001500,"started_at_run_us":1400000,"finished_at_run_us":1500000,"duration_us":100000,"fields":{"tt.kind":"request","tt.request_id":"route-8","tt.route":"/downstream","tt.outcome":"error"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000001410,"finished_at_unix_ms":1700000001485,"started_at_run_us":1410000,"finished_at_run_us":1485000,"duration_us":75000,"fields":{"tt.kind":"stage","tt.request_id":"route-8","tt.stage":"db","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000001401,"finished_at_unix_ms":1700000001406,"started_at_run_us":1401000,"finished_at_run_us":1406000,"duration_us":5000,"fields":{"tt.kind":"queue","tt.request_id":"route-8","tt.queue":"permits","tt.depth_at_start":5}}}
