+++ tailtriage-tracing/tests/fixtures/equivalence/semantic_retention_limits.jsonl
@@ -1,0 +1,15 @@
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000000,"finished_at_unix_ms":1700000000080,"started_at_run_us":0,"finished_at_run_us":80000,"duration_us":80000,"fields":{"tt.kind":"request","tt.request_id":"limit-1","tt.route":"/limited","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000030,"finished_at_unix_ms":1700000000050,"started_at_run_us":30000,"finished_at_run_us":50000,"duration_us":20000,"fields":{"tt.kind":"stage","tt.request_id":"limit-1","tt.stage":"stage-1","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000001,"finished_at_unix_ms":1700000000031,"started_at_run_us":1000,"finished_at_run_us":31000,"duration_us":30000,"fields":{"tt.kind":"queue","tt.request_id":"limit-1","tt.queue":"queue-1","tt.depth_at_start":1}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000100,"finished_at_unix_ms":1700000000180,"started_at_run_us":100000,"finished_at_run_us":180000,"duration_us":80000,"fields":{"tt.kind":"request","tt.request_id":"limit-2","tt.route":"/limited","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000130,"finished_at_unix_ms":1700000000150,"started_at_run_us":130000,"finished_at_run_us":150000,"duration_us":20000,"fields":{"tt.kind":"stage","tt.request_id":"limit-2","tt.stage":"stage-2","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000101,"finished_at_unix_ms":1700000000131,"started_at_run_us":101000,"finished_at_run_us":131000,"duration_us":30000,"fields":{"tt.kind":"queue","tt.request_id":"limit-2","tt.queue":"queue-2","tt.depth_at_start":2}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000200,"finished_at_unix_ms":1700000000280,"started_at_run_us":200000,"finished_at_run_us":280000,"duration_us":80000,"fields":{"tt.kind":"request","tt.request_id":"limit-3","tt.route":"/limited","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000230,"finished_at_unix_ms":1700000000250,"started_at_run_us":230000,"finished_at_run_us":250000,"duration_us":20000,"fields":{"tt.kind":"stage","tt.request_id":"limit-3","tt.stage":"stage-3","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000201,"finished_at_unix_ms":1700000000231,"started_at_run_us":201000,"finished_at_run_us":231000,"duration_us":30000,"fields":{"tt.kind":"queue","tt.request_id":"limit-3","tt.queue":"queue-3","tt.depth_at_start":3}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000300,"finished_at_unix_ms":1700000000380,"started_at_run_us":300000,"finished_at_run_us":380000,"duration_us":80000,"fields":{"tt.kind":"request","tt.request_id":"limit-4","tt.route":"/limited","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000330,"finished_at_unix_ms":1700000000350,"started_at_run_us":330000,"finished_at_run_us":350000,"duration_us":20000,"fields":{"tt.kind":"stage","tt.request_id":"limit-4","tt.stage":"stage-4","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000301,"finished_at_unix_ms":1700000000331,"started_at_run_us":301000,"finished_at_run_us":331000,"duration_us":30000,"fields":{"tt.kind":"queue","tt.request_id":"limit-4","tt.queue":"queue-4","tt.depth_at_start":4}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"request","started_at_unix_ms":1700000000400,"finished_at_unix_ms":1700000000480,"started_at_run_us":400000,"finished_at_run_us":480000,"duration_us":80000,"fields":{"tt.kind":"request","tt.request_id":"limit-5","tt.route":"/limited","tt.outcome":"ok"}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"stage","started_at_unix_ms":1700000000430,"finished_at_unix_ms":1700000000450,"started_at_run_us":430000,"finished_at_run_us":450000,"duration_us":20000,"fields":{"tt.kind":"stage","tt.request_id":"limit-5","tt.stage":"stage-5","tt.success":true}}}
+{"format":"tailtriage.tracing-span.v1","span":{"name":"queue","started_at_unix_ms":1700000000401,"finished_at_unix_ms":1700000000431,"started_at_run_us":401000,"finished_at_run_us":431000,"duration_us":30000,"fields":{"tt.kind":"queue","tt.request_id":"limit-5","tt.queue":"queue-5","tt.depth_at_start":5}}}
