+++ tailtriage-tracing/tests/support/mod.rs
@@ -1,1 +1,2 @@
 pub mod equivalence_harness;
+pub mod live_harness;
